# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/",
            "description": "Test API Server"
        }
    ],
    "paths": {
        "/categorias": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "operationId": "66f747250d1c2a1a99b78b5ffc76ad86",
                "responses": {
                    "200": {
                        "description": "Display a listing of categories."
                    }
                }
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "operationId": "5262fac7d8d2fea2144b3d9ef99c6635",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "descricao"
                                ],
                                "properties": {
                                    "descricao": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Store new category"
                    }
                }
            }
        },
        "/categorias/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "operationId": "c850dd742053be302993ad6e3c2f9f2b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "category id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Display a category by id."
                    }
                }
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "operationId": "32680efa60b9809623c4f4ed282d0c5b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "descricao"
                                ],
                                "properties": {
                                    "descricao": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update a category by id."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "operationId": "5c588af090fa094f2c5794ed8ce16aaa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete a category by id."
                    }
                }
            }
        },
        "/produtos": {
            "get": {
                "tags": [
                    "Products"
                ],
                "operationId": "a28d49d6336a249ac3bf0fdb3deb516f",
                "responses": {
                    "200": {
                        "description": "Display a listing of products."
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "operationId": "4ef2f8bfa3024fa10e6e5e26f25a7dd0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "descricao",
                                    "dimensoes",
                                    "codigo",
                                    "referencia",
                                    "saldo_estoque",
                                    "preco",
                                    "categoria_id"
                                ],
                                "properties": {
                                    "descricao": {
                                        "type": "string"
                                    },
                                    "dimensoes": {
                                        "type": "string"
                                    },
                                    "codigo": {
                                        "type": "string"
                                    },
                                    "referencia": {
                                        "type": "string"
                                    },
                                    "saldo_estoque": {
                                        "type": "integer"
                                    },
                                    "preco": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "categoria_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Store new product"
                    }
                }
            }
        },
        "/produtos/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "operationId": "dd22a170cd1a1cd4de23ce8cbd7e6b77",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Display a product by id."
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "operationId": "f51301fab977d8eb15f4969cdd5775a9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "descricao",
                                    "dimensoes",
                                    "codigo",
                                    "referencia",
                                    "saldo_estoque",
                                    "preco",
                                    "categoria_id"
                                ],
                                "properties": {
                                    "descricao": {
                                        "type": "string"
                                    },
                                    "dimensoes": {
                                        "type": "string"
                                    },
                                    "codigo": {
                                        "type": "string"
                                    },
                                    "referencia": {
                                        "type": "string"
                                    },
                                    "saldo_estoque": {
                                        "type": "integer"
                                    },
                                    "preco": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "categoria_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update a product by id."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "operationId": "60a103a91426588fa605c381fe0c8180",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete a product by id."
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Projects",
            "description": "API Endpoints of Projects"
        }
    ]
}